backend/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docx2txt==0.8
et-xmlfile==1.1.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
lxml==4.9.3
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pdfminer.six==20231228
Pillow==10.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyMySQL==1.1.0
pypdf2==3.2.1
python-dateutil==2.8.2
python-docx==1.1.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
python-pptx==0.6.23
openpyxl==3.1.2
pypdf==4.0.1
pymupdf>=1.24.0
python-docx==1.1.0
langchain-openai==0.0.8
langchain-community==0.0.24
langchain-core==0.1.26
dashscope==1.15.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-dotenv==1.0.1
loguru==0.7.2
ragas==0.1.2
datasets==2.16.1
tiktoken==0.5.2
nest-asyncio==1.6.0
httpx==0.26.0
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
scikit-learn==1.4.0
numpy==1.26.4
md5-hash==0.1.0
